+++ src/telas/TelaPrincipal.form
@@ -86,9 +86,11 @@
               <Group type="103" groupAlignment="0" attributes="0">
                   <Group type="102" alignment="0" attributes="0">
                       <EmptySpace min="-2" pref="35" max="-2" attributes="0"/>
-                      <Group type="103" groupAlignment="1" attributes="0">
+                      <Group type="103" groupAlignment="1" max="-2" attributes="0">
                           <Component id="jScrollPane1" min="-2" pref="1065" max="-2" attributes="0"/>
-                          <Group type="102" alignment="1" attributes="0">
+                          <Group type="102" attributes="0">
+                              <Component id="jbAtualizar" min="-2" max="-2" attributes="0"/>
+                              <EmptySpace max="32767" attributes="0"/>
                               <Component id="jbExcluirLivro" min="-2" pref="150" max="-2" attributes="0"/>
                               <EmptySpace min="-2" pref="40" max="-2" attributes="0"/>
                               <Component id="jbCadastrar" min="-2" pref="150" max="-2" attributes="0"/>
@@ -116,6 +118,7 @@
                           <Component id="jbCadastrar" alignment="3" min="-2" max="-2" attributes="0"/>
                           <Component id="jbExcluirLivro" alignment="3" min="-2" max="-2" attributes="0"/>
                           <Component id="jbCadastrarCliente" alignment="3" min="-2" max="-2" attributes="0"/>
+                          <Component id="jbAtualizar" alignment="3" min="-2" max="-2" attributes="0"/>
                       </Group>
                       <EmptySpace max="32767" attributes="0"/>
                   </Group>
@@ -133,7 +136,7 @@
                 <Component class="javax.swing.JTable" name="jTable1">
                   <Properties>
                     <Property name="model" type="javax.swing.table.TableModel" editor="org.netbeans.modules.form.editors2.TableModelEditor">
-                      <Table columnCount="4" rowCount="4">
+                      <Table columnCount="4" rowCount="0">
                         <Column editable="true" title="T&#xed;tulo" type="java.lang.String"/>
                         <Column editable="true" title="C&#xf3;digo" type="java.lang.Integer"/>
                         <Column editable="true" title="Autor" type="java.lang.String"/>
@@ -200,6 +203,14 @@
               </Properties>
               <Events>
                 <EventHandler event="actionPerformed" listener="java.awt.event.ActionListener" parameters="java.awt.event.ActionEvent" handler="jbCadastrarClienteActionPerformed"/>
+              </Events>
+            </Component>
+            <Component class="javax.swing.JButton" name="jbAtualizar">
+              <Properties>
+                <Property name="text" type="java.lang.String" value="Atualizar Lista"/>
+              </Properties>
+              <Events>
+                <EventHandler event="actionPerformed" listener="java.awt.event.ActionListener" parameters="java.awt.event.ActionEvent" handler="jbAtualizarActionPerformed"/>
               </Events>
             </Component>
           </SubComponents>
